Correct nullable equality and allow it in joins
The instance for PGEq ('Nullable a) was wrong. It would compile to SQL
such as

   a = b OR (a IS NULL AND b IS NULL)

But if a or b is null, then the whole result will be null, not a
boolean. Instead, this instance has been removed, and replaced with a
new ?= operator.

Joins and restriction now take ('Nullable 'PGBoolean) as predicates.
This is for compatibility with (?=), but more importantly - it produces
better query plans. Checking the inputs for null and dealing with them
specifically might give nice semantics in Haskell, but it increases the
amount of work PostgreSQL has to do.