Correct gdistributeMaybe for nullable fields
gdistributeMaybe is used to turn a record of Maybe fields into a Maybe
record. If a field in a record is normally nullable (regardless of the
join type) then we should always succeed to distribute Maybe over this
field. However, the previous implementation depended on the value in
that field, which meant nulls would cause left joins to fail to match.